refactor: separate editor windows
